Working on polyphony

+++ include/states.hpp
@@ -3,7 +3,7 @@
 namespace patch_magic
 {
     
-struct sine_state { float phase_ = 0.0f; };
+struct sine_state { double phase_ = 0.0f; };
 
 struct stateless {};
